+++ apps/web/src/app/(app)/dashboard/page.tsx
@@ -1,6 +1,6 @@
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 import { getQueryClient } from "@/lib/get-query-client";
-import { tripKeys } from "@/hooks/use-trips";
+import { tripKeys } from "@/hooks/trip-queries";
 import { serverApiRequest } from "@/lib/server-api";
 import { DashboardContent } from "./dashboard-content";
 import type { GetTripsResponse } from "@tripful/shared/types";
